Bug fixes, initial BaseLiteral support
By implementing and using flatMapWithLoc it is now possible to report parser errors up the responder chain. You have to be carefull to not backtrack at the wrong position, because otherwise the parser errors will be converted from an arresting failures to a epsilon failures and the parser will proceed without emitting the error.
Initial support for BaseLiterals was added.