Add inheritance to GenerateDatalog
Each time an instance of a subclass is created, an instance of the parent class with the same uri will be created. This solution isn't very nice right now, since it requires passing around an empty MockURI if no parent class is available and a child URI will be stored twice in the "this" field and the "childURI" field.