Introduce mutation for variables and refactor the objectoriented transformations
When a variable is mutated, a new variable with the name of the old one is created using gensym.fresh. When a VarReadExpr is encountered, the last defined freshly generated variable name is used. Since variable assignment is only possible inside the scope of a function, it should be save to implement it this way. For immutable variables the old behaviour will be used.

The FieldTransformation and AllocTransformation have been refactored to use a common CounterTransformer.