Fix object creation inside fold
When creating a new object in scala and returning it back to datalog (e.g. Sum over objects of type Num), a new allocId must be assigned. This commit checks if the allocId is None and uses the datalog allocId in this case. In all other cases the allocId from the scala object is used to recreate the datalog object"