Fix uncoalescing of new objects created in an aggregation method
We need to override equals and hash for all scala generated classes. Uncoalesing and coalesing are both input relations. That means they might generate different scala objects when calculating the input. Nevertheless these objects must be considered equal if their attributes match. We can not use a case class, since we might inherit from a class, which is not possible when using a case class. See: SetFoldSumObject for a test where these overrrides are required.