Fix timestamp for case classes constructor
All classes use a fixed timestamp of 0 for default initialization of fields and for the primary constructor. The timestamp root will start at 1. This prevents increasing the timestamp in case of creating a new instance of a case class. In the future this can be removed, by eliminating timestamps completely for immutable fields.