Add WildCard and WeakNot
To support WildCards a new open Datatype Arg is introduced in the BaseIR. This type is either a WildCard or a Term. Arg can be placed in Calls and Call-like atoms, such as aggregations. The not lowering will introduce wildcards for binding variables in case of a weaknot. Aggregations introduce Arg by a special AggregateColumn case.