Fix Impure does not work with disjunction
We need to apply Impurity before we apply Disjunction, since the Disjunction lowering can not know how to rewrite the disjunction, if it is nested inside a Impurity.
The Impurity lowering on the other hand must be adapted to only insert equality constraints to the output parameter in a body of a relation, not inside a disjunction.