Fix impure lowering to correctly handle nested disjunctions
The impure lowering did not handle disjunctions nested in a body correctly. I described the problem in detail in the impure lowering file under the note 5. While the idea is simple, the implementation is not. The implementation is really messy. If anybody has a nicer approach to achive the same result, feel more than invited to change this implementation.

That said, with this new implementation the DependencyAnalysis (without mono.Set) is now terminating.